+++ tests/testthat/test_loglikelihood.R
@@ -34,7 +34,7 @@ test_that("log-likelihoods match for two-level model", {
   lfun_AGQ_pars <- lfun_AGQ(pars)
   lfun_SR_pars <- lfun_SR(pars)
   lfun_IS_pars <- lfun_IS(pars)
-  
+
   lfun_L2 <- find_lfun_glmm(modfr, method = "Laplace", control = list(order = 2))
   lfun_L2_pars <- lfun_L2(pars)
 
@@ -43,4 +43,18 @@ test_that("log-likelihoods match for two-level model", {
   expect_true(abs(lfun_AGQ_pars - lfun_IS_pars) < 0.1)
   expect_true(abs(lfun_L2_pars - lfun_Laplace_pars) > abs(lfun_L2_pars - lfun_AGQ_pars))
 
+})
+
+test_that("All methods work with a single random effect",{
+  y <- c(rep(1, 10), rep(0, 10))
+  i <- rep(1, 20)
+  modfr <- find_modfr_glmm(y ~ (1 | i), family = "binomial")
+  lfun_L_2 <- find_lfun_glmm(modfr, method = "Laplace", control = list(order = 2))
+  lfun_L_2(c(1, 1))
+  lfun_AGQ_5 <- find_lfun_glmm(modfr, method = "AGQ", control = list(nAGQ = 5))
+  lfun_AGQ_5(c(1, 1))
+  lfun_SR_3 <- find_lfun_glmm(modfr, method = "SR", control = list(nSL = 3))
+  lfun_SR_3(c(1, 1))
+  lfun_IS_100 <- find_lfun_glmm(modfr, method = "IS", control = list(nIS = 100))
+  lfun_IS_100(c(1, 1))
 })